Further progress toward a working topic trie.
There is a major flaw in the classic trie implementation: a trie "split" (e.g.
expanding '.'->'foo.bar.baz' to '.'->'foo'->'bar.baz') requires re-aligning the
contexts for the entire subtree.  This means that a "split" is insanely
inefficient.

The question is whether trie walks with a fully expanded trie (e.g. each topic
token gets its own node in the trie) or a high split cost is more of a
bottleneck.

I'm tempted to say "leave it as it is" for now and see if its a problem later.
The reason for this is that I expect trie splits to be relatively uncommon... we
are optimizing for publication after all, not subscription, and this is
subscription-related overhead, not publication-related.